script load retry and resin logo loading spinner
* incorporated `requirejs` to retry javascript loading in the frontend
* added resin logo for spinning loader

There are two loader stages - one for the GUI, one for the Frontend.
It's not easy to hand over between them, so doing two spinner logos
(that is one for each), and spin them opposite direction to make
look less horrible (ie. instead of spin-jerk-spin, it's more like
spin-then-spin-backwards).

Change-type: minor